requirements.txt
flask
flask-socketio
opencv-python
numpy
RPi.GPIO
adafruit-circuitpython-motorkit
